+++ var/spack/repos/builtin/packages/of-precice/package.py
@@ -8,7 +8,7 @@ import os
 import llnl.util.tty as tty
 
 from spack import *
-from spack.pkg.builtin.openfoam_com import add_extra_files
+from spack.pkg.builtin.openfoam import add_extra_files
 
 
 class OfPrecice(Package):
