--- var/spack/repos/builtin/packages/openfoam-com/common/README
@@ -1,2 +1,0 @@
-Some helper tools for packaging applications/libraries dependent on an
-openfoam provider.
