+++ var/spack/repos/builtin/packages/openfoam/1806-have-kahip.patch
