+++ var/spack/repos/builtin/packages/openfoam/common/README
@@ -1,0 +1,1 @@
+Helper tools for packaging applications/libraries dependent on OpenFOAM.
